fix Zseg signedness

+++ prboom2/src/p_setup.c
@@ -2056,8 +2056,8 @@ static void P_LoadLineDefs (int lump)
       ld->special_args[4] = mld->arg5;
       ld->v1 = &vertexes[(unsigned short)LittleShort(mld->v1)];
       ld->v2 = &vertexes[(unsigned short)LittleShort(mld->v2)];
-      ld->sidenum[0] = LittleShort(mld->sidenum[0]);
-      ld->sidenum[1] = LittleShort(mld->sidenum[1]);
+      ld->sidenum[0] = (unsigned short)LittleShort(mld->sidenum[0]);
+      ld->sidenum[1] = (unsigned short)LittleShort(mld->sidenum[1]);
       P_SetLineID(ld);
     }
     else
@@ -2074,8 +2074,8 @@ static void P_LoadLineDefs (int lump)
       ld->special_args[4] = 0;
       ld->v1 = &vertexes[(unsigned short)LittleShort(mld->v1)];
       ld->v2 = &vertexes[(unsigned short)LittleShort(mld->v2)];
-      ld->sidenum[0] = LittleShort(mld->sidenum[0]);
-      ld->sidenum[1] = LittleShort(mld->sidenum[1]);
+      ld->sidenum[0] = (unsigned short)LittleShort(mld->sidenum[0]);
+      ld->sidenum[1] = (unsigned short)LittleShort(mld->sidenum[1]);
     }
 
     if ((unsigned short) -1 == ld->sidenum[0])
